requirements.txt
# Google Cloud Services
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-cloud-aiplatform==1.43.0

# Data Processing
python-dotenv==1.0.1
requests==2.31.0

# Logging
coloredlogs==15.0.1

# Retry logic
tenacity==8.2.3
